requirements.txt
requests
fpdf2
Pillow